requirements.txt
South==0.7.4
djlime>=0.0.8
django-imagekit==2.0.1